+++ src/main/java/com/ticketscale/infrastructure/security/SecurityConfigurations.java
@@ -32,6 +32,7 @@ public class SecurityConfigurations {
                     req.requestMatchers(HttpMethod.POST, "/usuarios").permitAll();
                     req.requestMatchers(HttpMethod.POST, "/eventos").hasRole("ADMIN");
                     req.requestMatchers(HttpMethod.DELETE, "/eventos/**").hasRole("ADMIN");
+                    req.requestMatchers("/dashboard/**").hasRole("ADMIN");
                     req.anyRequest().authenticated();
                 })
                 .addFilterBefore(securityFilter, UsernamePasswordAuthenticationFilter.class)